Refactor server package to integrate new storage client and update authentication service
